update bookings endpoints to pass query parameter based on jobstatus for update and get bookings. Also add place to booking DTO response in place of of place_id previously returned

+++ src/main/java/com/hdpros/hdprosbackend/booking/dto/BookingDTO.java
@@ -19,7 +19,7 @@ public class BookingDTO {
 
     private String description;
 
-    private int placeId;
+    private Long placeId;
 
     private boolean jobStatus;
